test:修改react-native-modalbox测试demo
